Concepts Edit fertig

+++ db/migrate/20170204163158_create_concepts.rb
@@ -2,8 +2,8 @@ class CreateConcepts < ActiveRecord::Migration[5.0]
   def change
     create_table :concepts do |t|
       t.string :label
-      t.integer :x
-      t.integer :y
+      t.float :x
+      t.float :y
       t.integer :concept_map_id
 
       t.timestamps
